fastapi_backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
